Reference

Index

GpAbc.AbstractGPKernel

GpAbc.GPModel

GpAbc.GPModel

GpAbc.GPModel

GpAbc.GPModel

GpAbc.MaternArdKernel

GpAbc.MaternIsoKernel

GpAbc.SquaredExponentialArdKernel

GpAbc.SquaredExponentialIsoKernel

GpAbc.ExponentialArdKernel

GpAbc.ExponentialIsoKernel

GpAbc.covariance

GpAbc.covariance_diagonal

GpAbc.covariance_grad

GpAbc.covariance_training

GpAbc.get_hyperparameters_size

GpAbc.gp_loglikelihood

GpAbc.gp_loglikelihood

GpAbc.gp_loglikelihood_grad

GpAbc.gp_loglikelihood_log

GpAbc.gp_regression

GpAbc.gp_regression

GpAbc.gp_train

GpAbc.scaled_squared_distance

GpAbc.scaled_squared_distance_grad

GpAbc.set_hyperparameters

GpAbc

GpAbc.AbstractGPKernel — Type.

AbstractGPKernel

Abstract kernel type. User-defined kernels should derive from it.

Implementations have to provide methods for get_hyperparameters_size and covariance. Methods for covariance_training, covariance_diagonal and covariance_grad are optional.

source

GpAbc.GPModel — Type.

GPModel

The main type that is used by most functions withing the package.

All data matrices are row-major.

Fields

kernel::AbstractGPKernel: the kernel

gp_training_x::AbstractArray{Float64, 2}: training x. Size: $n \times d$.

gp_training_y::AbstractArray{Float64, 2}: training y. Size: $n \times 1$.

gp_test_x::AbstractArray{Float64, 2}: test x. Size: $m \times d$.

gp_hyperparameters::AbstractArray{Float64, 1}: kernel hyperparameters, followed by standard deviation of intrinsic noise $\sigma_n$, which is always the last element in the array.

cache::HPOptimisationCache: cache of matrices that can be re-used between calls to gp_loglikelihood and gp_loglikelihood_grad

source

GpAbc.GPModel — Type.

GPModel(training_x::Union{AbstractArray{Float64, 2}, AbstractArray{Float64, 1}},
        training_y::Union{AbstractArray{Float64, 2}, AbstractArray{Float64, 1}},
        kernel::AbstractGPKernel
        [,test_x::Union{AbstractArray{Float64, 2}, AbstractArray{Float64, 1}}=zeros(0,0)])

Constructor of GPModel that allows the kernel to be specified. Arguments that are passed as 1-d vectors will be reshaped into 2-d.

source

GpAbc.GPModel — Type.

GPModel(training_x::Union{AbstractArray{Float64, 2}, AbstractArray{Float64, 1}},
        training_y::Union{AbstractArray{Float64, 2}, AbstractArray{Float64, 1}}
        [,test_x::Union{AbstractArray{Float64, 2}, AbstractArray{Float64, 1}}=zeros(0,0)])

Default constructor of GPModel, that will use SquaredExponentialIsoKernel. Arguments that are passed as 1-d vectors will be reshaped into 2-d.

source

GpAbc.GPModel — Method.

GPModel(;training_x::Union{AbstractArray{Float64, 2}, AbstractArray{Float64, 1}}=zeros(0,0),
    training_y::Union{AbstractArray{Float64, 2}, AbstractArray{Float64, 1}}=zeros(0,0),
    test_x::Union{AbstractArray{Float64, 2}, AbstractArray{Float64, 1}}=zeros(0,0),
    kernel::AbstractGPKernel=SquaredExponentialIsoKernel(),
    gp_hyperparameters::AbstractArray{Float64, 1}=Array{Float64}(0))

Constructor of GPModel with explicit arguments. Arguments that are passed as 1-d vectors will be reshaped into 2-d.

source

GpAbc.MaternArdKernel — Type.

MaternArdKernel <: AbstractGPKernel

Matérn kernel with distinct length scale for each dimention, $l_k$. Parameter $\nu$ (nu) is passed in constructor. Currently, only values of $\nu=1$, $\nu=3$ and $\nu=5$ are supported.

\[\begin{aligned} K_{\nu=1}(r) &= \sigma_f^2e^{-\sqrt{r}}\\ K_{\nu=3}(r) &= \sigma_f^2(1 + \sqrt{3r})e^{-\sqrt{3r}}\\ K_{\nu=5}(r) &= \sigma_f^2(1 + \sqrt{3r} + \frac{5}{3}r)e^{-\sqrt{5r}}\\ r_{ij} &= \sum_{k=1}^d\frac{(x_{ik}-z_{jk})^2}{l_k^2} \end{aligned}\]

$r_{ij}$ are computed by scaled_squared_distance

Hyperparameters

The length of hyperparameters array for this kernel depends on the dimensionality of the data. Assuming each data point is a vector in a $d$-dimensional space, this kernel needs $d+1$ hyperparameters, in the following order:

$\sigma_f$: the signal standard deviation

$l_1, \ldots, l_d$: the length scales for each dimension

source

GpAbc.MaternIsoKernel — Type.

MaternIsoKernel <: AbstractGPKernel

Matérn kernel with uniform length scale across all dimensions, $l$. Parameter $\nu$ (nu) is passed in constructor. Currently, only values of $\nu=1$, $\nu=3$ and $\nu=5$ are supported.

\[\begin{aligned} K_{\nu=1}(r) &= \sigma_f^2e^{-\sqrt{r}}\\ K_{\nu=3}(r) &= \sigma_f^2(1 + \sqrt{3r})e^{-\sqrt{3r}}\\ K_{\nu=5}(r) &= \sigma_f^2(1 + \sqrt{3r} + \frac{5}{3}r)e^{-\sqrt{5r}}\\ r_{ij} &= \sum_{k=1}^d\frac{(x_{ik}-z_{jk})^2}{l^2} \end{aligned}\]

$r_{ij}$ are computed by scaled_squared_distance

Hyperparameters

Hyperparameters vector for this kernel must contain two elements, in the following order:

$\sigma_f$: the signal standard deviation

$l$: the length scale

source

GpAbc.SquaredExponentialArdKernel — Type.

SquaredExponentialArdKernel <: AbstractGPKernel

Squared exponential kernel with distinct length scale for each dimention, $l_k$.

\[\begin{aligned} K(r) & = \sigma_f^2 e^{-r/2} \\ r_{ij} & = \sum_{k=1}^d\frac{(x_{ik}-z_{jk})^2}{l_k^2} \end{aligned}\]

$r_{ij}$ are computed by scaled_squared_distance

Hyperparameters

The length of hyperparameters array for this kernel depends on the dimensionality of the data. Assuming each data point is a vector in a $d$-dimensional space, this kernel needs $d+1$ hyperparameters, in the following order:

$\sigma_f$: the signal standard deviation

$l_1, \ldots, l_d$: the length scales for each dimension

source

GpAbc.SquaredExponentialIsoKernel — Type.

SquaredExponentialIsoKernel <: AbstractGPKernel

Squared exponential kernel with uniform length scale across all dimensions, $l$.

\[\begin{aligned} K(r) & = \sigma_f^2 e^{-r/2} \\ r_{ij} & = \sum_{k=1}^d\frac{(x_{ik}-z_{jk})^2}{l^2} \end{aligned}\]

$r_{ij}$ are computed by scaled_squared_distance

Hyperparameters

Hyperparameters vector for this kernel must contain two elements, in the following order:

$\sigma_f$: the signal standard deviation

$l$: the length scale

source

GpAbc.ExponentialArdKernel — Method.

ExponentialArdKernel

Alias for MaternArdKernel(1)

source

GpAbc.ExponentialIsoKernel — Method.

ExponentialIsoKernel

Alias for MaternIsoKernel(1)

source

GpAbc.covariance — Method.

covariance(ker::AbstractGPKernel, log_theta::AbstractArray{Float64, 1},
    x::AbstractArray{Float64, 2}, z::AbstractArray{Float64, 2})

Return the covariance matrix. Should be overridden by kernel implementations.

Arguments

ker: The kernel object. Implementations must override with their own subtype.

log_theta: natural logarithm of hyperparameters.

x, z: Input data, reshaped into 2-d arrays. x must have dimensions $n \times d$; z must have dimensions $m \times d$.

Return

The covariance matrix, of size $n \times m$.

source

GpAbc.covariance_diagonal — Method.

covariance_diagonal(ker::AbstractGPKernel, log_theta::AbstractArray{Float64, 1},
    x::AbstractArray{Float64, 2})

This is a speedup version of covariance, which is invoked if the caller is not interested in the entire covariance matrix, but only needs the variance, i.e. the diagonal of the covariance matrix.

Default method just returns diag(covariance(...)), with x === z. Kernel implementations can optionally override it to achieve betrer performance, by not computing the non diagonal elements of covariance matrix.

See covariance for description of arguments.

Return

The 1-d array of variances, of size size(x, 1).

source

GpAbc.covariance_grad — Method.

covariance_grad(ker::AbstractGPKernel, log_theta::AbstractArray{Float64, 1},
    x::AbstractArray{Float64, 2}, R::AbstractArray{Float64, 2})

Return the gradient of the covariance function with respect to logarigthms of hyperparameters, based on the provided direction matrix.

This function can be optionally overridden by kernel implementations. If the gradient function is not provided, gp_train will fail back to NelderMead algorithm by default.

Arguments

ker: The kernel object. Implementations must override with their own subtype.

log_theta: natural logarithm of hyperparameters

x: Training data, reshaped into a 2-d array. x must have dimensions $n \times d$.

R the directional matrix, $n \times n$

\[R = \frac{1}{\sigma_n^2}(\alpha * \alpha^T - K^{-1}); \alpha = K^{-1}y\]

Return

A vector of size length(log_theta), whose $j$'th element is equal to

\[tr(R \frac{\partial K}{\partial \eta_j})\]

source

GpAbc.covariance_training — Method.

covariance_training(ker::AbstractGPKernel, log_theta::AbstractArray{Float64, 1},
    training_x::AbstractArray{Float64, 2})

This is a speedup version of covariance, which is only called during traing sequence. Intermediate matrices computed in this function for particular hyperparameters can be cached and reused subsequently, either in this function or in covariance_grad

Default method just delegates to covariance with x === z. Kernel implementations can optionally override it for betrer performance.

See covariance for description of arguments and return values.

source

GpAbc.get_hyperparameters_size — Method.

get_hyperparameters_size(kernel::AbstractGPKernel, training_data::AbstractArray{Float64, 2})

Return the number of hyperparameters for used by this kernel on this training data set. Should be overridden by kernel implementations.

source

GpAbc.gp_loglikelihood — Method.

gp_loglikelihood(theta::AbstractArray{Float64, 1}, gpm::GPModel)

source

GpAbc.gp_loglikelihood — Method.

gp_loglikelihood(gpm::GPModel)

Compute the log likelihood function, based on the kernel and training data specified in gpm.

\[log p(y \vert X, \theta) = - \frac{1}{2}(y^TK^{-1}y + log \vert K \vert + n log 2 \pi)\]

source

GpAbc.gp_loglikelihood_grad — Method.

gp_loglikelihood_grad(theta::AbstractArray{Float64, 1}, gpem::GPModel)

Gradient of the log likelihood function (gp_loglikelihood_log) with respect to logged hyperparameters.

source

GpAbc.gp_loglikelihood_log — Method.

gp_loglikelihood_log(theta::AbstractArray{Float64, 1}, gpm::GPModel)

Log likelihood function with log hyperparameters. This is the target function of the hyperparameters optimisation procedure. Its gradient is coputed by gp_loglikelihood_grad.

source

GpAbc.gp_regression — Method.

gp_regression(gpm::GPModel; <optional keyword arguments>)

Run the Gaussian Process Regression.

Arguments

gpm: the GPModel, that contains the training data (x and y), the kernel, the hyperparameters and the test data for running the regression.

test_x: if specified, overrides the test x in gpm. Size $m \times d$.

log_level::Int (optional): log level. Default is 0, which is no logging at all. 1 makes gp_regression print basic information to standard output.

full_covariance_matrix::Bool (optional): whether we need the full covariance matrix, or just the variance vector. Defaults to false (i.e. just the variance).

batch_size::Int (optional): If full_covariance_matrix is set to false, then the mean and variance vectors will be computed in batches of this size, to avoid allocating huge matrices. Defaults to 1000.

observation_noise::Bool (optional): whether the observation noise (with variance $\sigma_n^2$) should be included in the output variance. Defaults to true.

Return

A tuple of (mean, var). mean is a mean vector of the output multivariate Normal distribution, of size $m$. var is either the covariance matrix of size $m \times m$, or a variance vector of size $m$, depending on full_covariance_matrix flag.

source

GpAbc.gp_regression — Method.

gp_regression(test_x::Union{AbstractArray{Float64, 1}, AbstractArray{Float64, 2}},
    gpem::GPModel; <optional keyword arguments>)

source

GpAbc.gp_train — Method.

gp_train(gpm::GPModel; <optional keyword arguments>)

Find Maximum Likelihood Estimate of Gaussian Process hyperparameters by maximising gp_loglikelihood, using Optim package.

Arguments

gpm: the GPModel, that contains the training data (x and y), the kernel and the starting hyperparameters that will be used for optimisation.

optimisation_solver_type::Type{<:Optim.Optimizer} (optional): the solver to use. If not given, then ConjugateGradient will be used for kernels that have gradient implementation, and NelderMead will be used for those that don't.

hp_lower::AbstractArray{Float64, 1} (optional): the lower boundary for box optimisation. Defaults to $e^{-10}$ for all hyperparameters.

hp_upper::AbstractArray{Float64, 1} (optional): the upper boundary for box optimisation. Defaults to $e^{10}$ for all hyperparameters.

log_level::Int (optional): log level. Default is 0, which is no logging at all. 1 makes gp_train print basic information to standard output. 2 switches Optim logging on, in addition to 1.

Return

The list of all hyperparameters, including the standard deviation of the measurement noise $\sigma_n$. Note that after this function returns, the hyperparameters of gpm will be set to the optimised value, and there is no need to call set_hyperparameters once again.

source

GpAbc.scaled_squared_distance — Method.

scaled_squared_distance(log_ell::AbstractArray{Float64, 1},
    x::AbstractArray{Float64, 2}, z::AbstractArray{Float64, 2})

Compute the scaled squared distance between x and z:

\[r_{ij} = \sum_{k=1}^d\frac{(x_{ik}-z_{jk})^2}{l_k^2}\]

The gradient of this function with respect to length scale hyperparameter(s) is returned by scaled_squared_distance_grad.

Arguments

x, z: Input data, reshaped into 2-d arrays. x must have dimensions $n \times d$; z must have dimensions $m \times d$.

log_ell: logarithm of length scale(s). Can either be an array of size one (isotropic), or an array of size d (ARD)

Return

An $n \times m$ matrix of scaled squared distances

source

GpAbc.scaled_squared_distance_grad — Method.

scaled_squared_distance_grad(log_ell::AbstractArray{Float64, 1},
    x::AbstractArray{Float64, 2}, z::AbstractArray{Float64, 2}, R::AbstractArray{Float64, 2})

Return the gradient of the scaled_squared_distance function with respect to logarigthms of length scales, based on the provided direction matrix.

Arguments

x, z: Input data, reshaped into 2-d arrays. x must have dimensions $n \times d$; z must have dimensions $m \times d$.

log_ell: logarithm of length scale(s). Can either be an array of size one (isotropic), or an array of size d (ARD)

R the direction matrix, $n \times m$. This can be used to compute the gradient of a function that depends on scaled_squared_distance via the chain rule.

Return

A vector of size length(log_ell), whose $k$'th element is equal to

\[\text{tr}(R \frac{\partial K}{\partial l_k})\]

source

GpAbc.set_hyperparameters — Method.

set_hyperparameters(gpm::GPModel, hypers::AbstractArray{Float64, 1})

Set the hyperparameters of the GPModel

source