ABC Advanced

ABC Advanced Reference

Advanced aspects of GpABC support for parameter estimation with Approximate Bayesian Computation.

Index

GpABC.AbstractEmulatedParticleSelection

GpABC.AbstractEmulatorRetraining

GpABC.AbstractEmulatorTraining

GpABC.DefaultEmulatorTraining

GpABC.IncrementalRetraining

GpABC.MeanEmulatedParticleSelection

GpABC.MeanVarEmulatedParticleSelection

GpABC.NoopRetraining

GpABC.PreviousPopulationRetraining

GpABC.PreviousPopulationThresholdRetraining

GpABC.abc_retrain_emulator

GpABC.abc_select_emulated_particles

GpABC.train_emulator

Types and Functions

GpABC.AbstractEmulatorTraining — Type.

AbstractEmulatorTraining

Subtypes of this abstract type control how the emulator is trained for emulation-based ABC algorithms (rejection and SMC). At the moment, only DefaultEmulatorTraining is shipped. Custom emulator training procedure can be implemented by creating new subtypes of this type and overriding train_emulator for them.

A typical use case would be trying to control the behaviour of gp_train more tightly, or not using it altogeather (e.g. using another optimisation package).

source

GpABC.DefaultEmulatorTraining — Type.

DefaultEmulatorTraining <: AbstractEmulatorTraining

Fields

kernel::AbstractGPKernel: the kernel (AbstractGPKernel) that will be used with the Gaussian Process (GPModel). Defaults to SquaredExponentialArdKernel.

train_emulator method with this argument type calls gp_train with default arguments.

source

GpABC.AbstractEmulatedParticleSelection — Type.

AbstractEmulatedParticleSelection

Subtypes of this type control the criteria that determine what particles are included in the posterior for emulation-based ABC. Custom strategies can be implemented by creating new subtypes of this type and overriding abc_select_emulated_particles for them.

Two implementations are shipped:

MeanEmulatedParticleSelection

MeanVarEmulatedParticleSelection

source

GpABC.MeanVarEmulatedParticleSelection — Type.

MeanVarEmulatedParticleSelection <: AbstractEmulatedParticleSelection

When this strategy is used, the particles for which both mean and standard deviation returned by gp_regression is below the ABC threshold are included in the posterior.

Fields

variance_threshold_factor: scaling factor, by which the ABC threshold is multiplied

before checking the standard deviation. Defaults to 1.0.

source

GpABC.MeanEmulatedParticleSelection — Type.

MeanEmulatedParticleSelection <: AbstractEmulatedParticleSelection

When this strategy is used, the particles for which only the mean value returned by gp_regression is below the ABC threshold are included in the posterior. Variance is not checked.

source

GpABC.AbstractEmulatorRetraining — Type.

AbstractEmulatorRetraining

Subtypes of this abstract type control the additional retraining procedure that may or may not be carried out before each iteration of emulation-based ABC. Custom strategies may be implemented by creating new subtypes of this type and new abc_retrain_emulator methods for them.

The following implementations are shipped:

IncrementalRetraining

PreviousPopulationRetraining

PreviousPopulationThresholdRetraining

NoopRetraining

source

GpABC.IncrementalRetraining — Type.

IncrementalRetraining <: AbstractEmulatorRetraining

This emulator retraining strategy samples extra particles from the previous population, and adds them to the set of design points that were used on the previous iteration. The new design points are filtered according to the threshold. The combined set of design points is used to train the new emulator.

Fields

design_points: number of design points to add on each iteration

max_simulations: maximum number of simulations to perform during re-trainging on each iteration

source

GpABC.PreviousPopulationRetraining — Type.

PreviousPopulationRetraining <: AbstractEmulatorRetraining

This emulator retraining strategy samples extra particles from the previous population, and uses them to re-train the emulator from scratch. No filtering of the new design points is performed. Design points from the previous iteration are discarded.

source

GpABC.PreviousPopulationThresholdRetraining — Type.

PreviousPopulationThresholdRetraining <: AbstractEmulatorRetraining

This emulator retraining strategy samples extra particles from the previous population, and uses them to re-train the emulator from scratch. Design points from the previous iteration are discarded. This strategy allows to control how many design points are sampled with distance to the reference data below the threshold.

Fields:

n_design_points: number of design points

n_below_threshold: number of design points below the threshold

max_iter: maximum number of simulations to perform on each re-training iteration

source

GpABC.NoopRetraining — Type.

NoopRetraining <: AbstractEmulatorRetraining

A sentinel retraining strategy that does not do anything. When used, the emulator is trained only once before

source

GpABC.train_emulator — Function.

train_emulator(training_x, training_y, emulator_training<:AbstractEmulatorTraining)

Train the emulator. For custom training procedure, a new subtype of AbstractEmulatorTraining should be created, and a method of this function should be created for it.

Returns

A trained emulator. Shipped implementations return a GPModel.

source

GpABC.abc_retrain_emulator — Function.

abc_retrain_emulator(
    gp_model,
    particle_sampling_function,
    epsilon,
    training_input::EmulatorTrainingInput,
    retraining_settings<:AbstractEmulatorRetraining
    )

Additional retraining procedure for the emulator that may or may not be executed before every iteration of emulation based ABC. Details of the procedure are determined by the subtype of AbstractEmulatorRetraining that is passed as an argument.

source

GpABC.abc_select_emulated_particles — Function.

abc_select_emulated_particles(emulator, particles, threshold, selection<:AbstractEmulatedParticleSelection)

Compute the approximate distances by running the regression-based emulator on the provided particles, and return the accepted particles. Acceptance strategy is determined by selection - subtype of AbstractEmulatedParticleSelection.

Return

A tuple of accepted distances, and indices of the accepted particles in the supplied particles array.

source